29.12.17

+++ Lab8/Lab8SLN/Lab8/Models/Earthquake.cs
@@ -10,7 +10,7 @@ namespace Lab8.Models
     {
         public int ID { get; set; }
         public int Rok { get; set; }
-        public int Siła { get; set; }
+        public decimal Siła { get; set; }
         public string Kraj { get; set; }
         public string Miejsce { get; set; }
